requirements.txt
peewee==3.17.7
Pyrogram==2.0.106
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
